% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adls.R
\name{adls}
\alias{adls}
\title{Create an adls object.}
\usage{
adls(base_url, token)
}
\arguments{
\item{base_url}{character or \code{url} object made using \code{\link[httr:parse_url]{httr::parse_url()}}, this is
the base URL for the datalake store. It may be convenient to use \code{\link[=adls_url]{adls_url()}} to construct this.}

\item{token}{\code{Token2.0} reference-class (R6) object from \strong{httr}.
It may be convenient to use \code{\link[AzureOAuth:oauth_token_azure]{AzureOAuth::oauth_token_azure()}} to construct this.}
}
\value{
An \code{adls} S3 object.
}
\description{
TODO: describe the components and why it makes things easier to keep them together.
}
\examples{
\dontrun{
  library("AzureOAuth")

  # create token (assumes Azure native app)
  token <- oauth_token_azure(
    tenant_id = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
    application_id = "ffffffff-gggg-hhhh-iiii-jjjjjjjjjjjj",
    name = "foo"
  )

  # create adls object with base-URL and token
  adls_example <- adls(
    base_url = adls_url("bar"),
    token = token
  )
}
}
\seealso{
\code{\link[AzureOAuth:oauth_token_azure]{AzureOAuth::oauth_token_azure()}},
\code{\link[AzureOAuth:oauth_service_token_azure]{AzureOAuth::oauth_service_token_azure()}},
}
