% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{adls_create}
\alias{adls_create}
\title{Create and write to a file.}
\usage{
adls_create(adls, file, path, overwrite = FALSE, permission = NULL)
}
\arguments{
\item{adls}{\code{adls} S3 object, generated using \code{\link[=adls]{adls()}}.}

\item{file}{\code{form_file} S3 object, representing the "thing" to be uploaded.
You can compose this using \code{\link[httr:upload_file]{httr::upload_file()}} or \code{\link[curl:form_file]{curl::form_file()}}.}

\item{path}{\code{character}, remote path, defined with respect to root-path
in \code{adls$base_url}. You may find \code{\link[=file.path]{file.path()}} to be useful to compose the path.}

\item{overwrite}{logical, indicating if existing (remote) file is to be overwritten.}

\item{permission}{\code{integer} (octal form), permission to assign to the object.}
}
\value{
A \code{logical} indicating success of the operation.
}
\description{
Create and write to a file.
}
\examples{
\dontrun{
  # create token (assumes Azure native app)
  token <- AzureOAuth::oauth_token_azure(
    tenant_id = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
    application_id = "ffffffff-gggg-hhhh-iiii-jjjjjjjjjjjj",
    name = "foo"
  )

  # create adls object with base-URL and token
  adls_example <- adls(
    base_url = adls_url("bar"),
    token = token
  )

  # for this example, write a temporary file
  temp_file <- tempfile(fileext = ".csv")
  write.csv(iris, file = temp_file)

  # upload file
  adls_create(
    adls_example,
    file = httr::file_upload(temp_file),
    path = "iris.csv"
  )
}
}
\seealso{
WebHDFS documentation for \href{http://hadoop.apache.org/docs/stable/hadoop-project-dist/hadoop-hdfs/WebHDFS.html#Open_and_Read_a_File}{"Open and Read a File"}
}
