% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_status.R
\name{adls_list_status}
\alias{adls_list_status}
\alias{adls_path_exists}
\title{List a directory.}
\usage{
adls_list_status(adls, path = NULL)

adls_path_exists(adls, path = NULL)
}
\arguments{
\item{adls}{\code{adls} S3 object, generated using \code{\link[=adls]{adls()}}.}

\item{path}{\code{character}, remote path, defined with respect to root-path
in \code{adls$base_url}. You may find \code{\link[=file.path]{file.path()}} to be useful to compose the path.}
}
\value{
If empty, returns \code{NULL}. Otherwise returns \code{tbl_df} with with directory listing.

\code{logical} indicating if the path exists
}
\description{
List a directory.
}
\examples{
\dontrun{
  # create token (assumes Azure native app)
  token <- AzureOAuth::oauth_token_azure(
    tenant_id = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
    application_id = "ffffffff-gggg-hhhh-iiii-jjjjjjjjjjjj",
    name = "foo"
  )

  # create adls object with base-URL and token
  adls_example <- adls(
    base_url = adls_url("bar"),
    token = token
  )

  # create a directory
  adls_mkdirs(adls_example, "baz")

  # list the root directory
  adls_liststatus(adls_example)
}
}
\seealso{
\code{\link[=adls]{adls()}}, \code{\link[=adls_url]{adls_url()}}
WebHDFS documentation for \href{http://hadoop.apache.org/docs/stable/hadoop-project-dist/hadoop-hdfs/WebHDFS.html#List_a_Directory}{"List a Directory"}
}
