% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkdirs.R
\name{adls_mkdirs}
\alias{adls_mkdirs}
\title{Make a directory.}
\usage{
adls_mkdirs(adls, path, permission = NULL)
}
\arguments{
\item{adls}{\code{adls} S3 object, generated using \code{\link[=adls]{adls()}}}

\item{path}{character, directory to create with respect to root -
\code{\link[=file.path]{file.path()}} can be used here}

\item{permission}{integer (octal), permissions to assign to the directory}
}
\value{
logical, indicating success
}
\description{
Make a directory.
}
\examples{
\dontrun{
  url <- adls_url("foo")
  adls_mkdirs("bar")
}
}
