% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{adls_open_to_raw}
\alias{adls_open_to_raw}
\alias{adls_open_to_text}
\alias{adls_open_to_file}
\title{Create and write to a file.}
\usage{
adls_open_to_raw(adls, path, offset = NULL, length = NULL)

adls_open_to_text(adls, path, encoding = "UTF-8", offset = NULL,
  length = NULL)

adls_open_to_file(adls, path, path_local, offset = NULL, length = NULL)
}
\arguments{
\item{adls}{\code{adls} S3 object, generated using \code{\link[=adls]{adls()}}.}

\item{path}{\code{character}, remote path, defined with respect to root-path
in \code{adls$base_url}. You may find \code{\link[=file.path]{file.path()}} to be useful to compose the path.}

\item{offset}{\code{numeric}, the position of the starting-byte.}

\item{length}{\code{numeric}, the number of bytes to be processed.}

\item{encoding}{\code{character}, encoding to use.}

\item{path_local}{\code{character}, path on local computer.}
}
\value{
\describe{
\item{\code{adls_open_to_raw()}}{A \code{raw} vector.}
}

\describe{
\item{\code{adls_open_to_text()}}{A single \code{character} string.}
}

\describe{
\item{\code{adls_open_to_file()}}{A \code{logical} indicating success of the operation.}
}
}
\description{
Create and write to a file.
}
\examples{
\dontrun{
  # create token (assumes Azure native app)
  token <- AzureOAuth::oauth_token_azure(
    tenant_id = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
    application_id = "ffffffff-gggg-hhhh-iiii-jjjjjjjjjjjj",
    name = "foo"
  )

  # create adls object with base-URL and token
  adls_example <- adls(
    base_url = adls_url("bar"),
    token = token
  )

  # for this example, write a temporary file
  temp_file <- tempfile(fileext = ".csv")
  write.csv(iris, file = temp_file)

  # upload file
  adls_create(
    adls_example,
    file = httr::file_upload(temp_file),
    path = "iris.csv"
  )

  # read file to a string
  adls_open_to_text(
    adls_example,
    path = "iris.csv"
  )
}
}
\seealso{
WebHDFS documentation for \href{http://hadoop.apache.org/docs/stable/hadoop-project-dist/hadoop-hdfs/WebHDFS.html#Open_and_Read_a_File}{"Open and Read a File"}
}
