% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adls.R
\name{adls_url}
\alias{adls_url}
\title{Base URL}
\usage{
adls_url(adls_name, path = NULL, domain_name = "azuredatalakestore.net")
}
\arguments{
\item{adls_name}{\code{character}, Azure Data Lake Store name}

\item{path}{\code{character}, path to add to default root.
You may find \code{\link[=file.path]{file.path()}} to be useful to compose the path.}

\item{domain_name}{\code{character}, domain name}
}
\value{
\code{url} object returned by \code{\link[httr:parse_url]{httr::parse_url()}}
}
\description{
This is a convenience function to return the base URL for the service.
}
\examples{
  adls_url("foo")
}
\seealso{
Azure Datalake Store documentation on
\href{https://docs.microsoft.com/en-us/rest/api/datalakestore/webhdfs-filesystem-apis#url-for-rest-calls}{"WebHDFS Filesystem API"}
}
