% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{POSIXct_from_ms}
\alias{POSIXct_from_ms}
\title{Makes a datetime object using milliseconds}
\usage{
POSIXct_from_ms(x, tz = "UTC")
}
\arguments{
\item{x}{numeric, number of milliseconds from UNIX epoch}

\item{tz}{character, Olson timezone}
}
\value{
POSIXct
}
\description{
Makes a datetime object using milliseconds
}
\keyword{internal}
