% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adls.R
\name{adls}
\alias{adls}
\title{Create an adls object}
\usage{
adls(url, token)
}
\arguments{
\item{url}{url}

\item{token}{token}
}
\value{
An \code{adls} S3 object.
}
\description{
Create an adls object
}
\examples{
  adls(url, token)
}
