% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{adls_create}
\alias{adls_create}
\title{Create and write to a file.}
\usage{
adls_create(adls, form_file, path, overwrite = FALSE, permission = NULL)
}
\arguments{
\item{adls}{\code{adls} S3 object, generated using \code{\link[=adls]{adls()}}}

\item{form_file}{form_file made using \code{\link[httr:upload_file]{httr::upload_file()}} or \code{\link[curl:form_file]{curl::form_file()}}}

\item{path}{character, directory to act on, defined with respect to path
defined in \code{adls$base_url} - \code{\link[=file.path]{file.path()}} can be used here}

\item{overwrite}{logical, indicating if existing file is to be overwritten}

\item{permission}{integer (octal form), permissions to assign to the directory}
}
\value{
A logical indicating success of the operation.
}
\description{
Create and write to a file.
}
\examples{
\dontrun{
  library("AzureOAuth")

  # create token (assumes Azure native app)
  token <- oauth_token_azure(
    tenant_id = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
    application_id = "ffffffff-gggg-hhhh-iiii-jjjjjjjjjjjj",
    name = "foo"
  )

  # create adls object with base-URL and token
  adls_example <- adls(
    base_url = adls_url("bar"),
    token = token
  )

  # upload file
  adls_mkdirs(adls_example, "baz")
}
}
\seealso{
\code{\link[=adls]{adls()}}, \code{\link[=adls_url]{adls_url()}}
WebHDFS documentation for \href{http://hadoop.apache.org/docs/stable/hadoop-project-dist/hadoop-hdfs/WebHDFS.html#Open_and_Read_a_File}{"Open and Read a File"}
}
