% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{adls_open_to_raw}
\alias{adls_open_to_raw}
\alias{adls_open_to_text}
\alias{adls_open_to_file}
\title{Create and write to a file.}
\usage{
adls_open_to_raw(adls, path, offset = NULL, length = NULL)

adls_open_to_text(adls, path, encoding = "UTF-8", offset = NULL,
  length = NULL)

adls_open_to_file(adls, path, path_local, offset = NULL, length = NULL)
}
\arguments{
\item{adls}{\code{adls} S3 object, generated using \code{\link[=adls]{adls()}}}

\item{path}{character, directory to act on, defined with respect to path
defined in \code{adls$base_url} - \code{\link[=file.path]{file.path()}} can be used here}

\item{offset}{numeric, the position of the starting-byte}

\item{length}{numeric, the number of bytes to be processed}

\item{encoding}{character, encoding to use}

\item{path_local}{character, path to file on local computer}
}
\value{
response object from \code{\link[httr:VERB]{httr::VERB()}}
}
\description{
Create and write to a file.
}
\examples{
\dontrun{
  library("AzureOAuth")

  # create token (assumes Azure native app)
  token <- oauth_token_azure(
    tenant_id = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
    application_id = "ffffffff-gggg-hhhh-iiii-jjjjjjjjjjjj",
    name = "foo"
  )

  # create adls object with base-URL and token
  adls_example <- adls(
    base_url = adls_url("bar"),
    token = token
  )

  # upload file
  adls_mkdirs(adls_example, "baz")
}
}
\seealso{
\code{\link[=adls]{adls()}}, \code{\link[=adls_url]{adls_url()}}
WebHDFS documentation for \href{http://hadoop.apache.org/docs/stable/hadoop-project-dist/hadoop-hdfs/WebHDFS.html#Open_and_Read_a_File}{"Open and Read a File"}
}
