% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{adls_rename}
\alias{adls_rename}
\title{Rename a file/directory.}
\usage{
adls_rename(adls, path, destination)
}
\arguments{
\item{adls}{\code{adls} S3 object, generated using \code{\link[=adls]{adls()}}}

\item{path}{character, directory to act on, defined with respect to path
defined in \code{adls$base_url} - \code{\link[=file.path]{file.path()}} can be used here}

\item{destination}{character, path for the destination with respect to path
defined in \code{adls$base_url} - \code{\link[=file.path]{file.path()}} can be used here}
}
\value{
A logical indicating the success of the operation.
}
\description{
Rename a file/directory.
}
\examples{
\dontrun{
  library("AzureOAuth")

  # create token (assumes Azure native app)
  token <- oauth_token_azure(
    tenant_id = "aaaaaaaa-bbbb-cccc-dddd-eeeeeeeeeeee",
    application_id = "ffffffff-gggg-hhhh-iiii-jjjjjjjjjjjj",
    name = "foo"
  )

  # create adls object with base-URL and token
  adls_example <- adls(
    base_url = adls_url("bar"),
    token = token
  )

  # create a directory
  adls_mkdirs(adls_example, "baz")

  # reanme the newly-created directory
  adls_rename(adls_example, "baz", "baz_new")
}
}
\seealso{
\code{\link[=adls]{adls()}}, \code{\link[=adls_url]{adls_url()}}
WebHDFS documentation for \href{http://hadoop.apache.org/docs/stable/hadoop-project-dist/hadoop-hdfs/WebHDFS.html#Delete_a_FileDirectory}{"Delete a File/Directory"}
}
