% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adls.R
\name{adls_url}
\alias{adls_url}
\title{Base URL}
\usage{
adls_url(adls_name, domain_name = "azuredatalakestore.net")
}
\arguments{
\item{adls_name}{character, Azure Data Lake Store name}

\item{domain}{character, domain name}
}
\value{
\code{url} object returned by \code{\link[httr:parse_url]{httr::parse_url()}}
}
\description{
This is a convenience function to return the base URL for the service.
}
\examples{
  adls_url("foo")
}
\keyword{internal}
